Recursively evaluate expanded variables
Not entirely sure about whether this is correct or the right approach; The unit test shows the issue: variables values could themselves be expressions which contain variables, and so on.

This particular solution seems to be a bit messy, and it has a side-effect in that it will also result in a variables whose value is a string containing only a number to be expanded into an integer. It also fails to catch recursive variables (i.e. variables that end up expanding into something containing the same variable name) and will fail with a RecursionError